Adds /LogFile parameter to DotCoverSettings Solves #1151
